+++ CHANGELOG.md
@@ -1,10 +1,9 @@
 # 1.1.0
-
 - Moved to Tools -> VRC Bulk Upload
 - Fix for having multiple active scenes
 - Now ignores disabled avatar descriptors 
 - Added a timestamp to successful builds
+- Added a saftey check to ensure vrcAvatar ID & pipeline ID match.
 
 # 1.0.0
-
 - first release